requirements/base.txt
arrow==0.8.0
binaryornot==0.4.0
chardet==2.3.0
click==6.6
cookiecutter==1.4.0
coverage==4.1
Django==1.9.7
django-bootstrap3==7.0.1
django-debug-toolbar==1.4
django-registration==2.1.2
factory-boy==2.7.0
fake-factory==0.5.8
future==0.15.2
ipaddress==1.0.16
Jinja2==2.8
jinja2-time==0.2.0
MarkupSafe==0.23
poyo==0.4.0
python-dateutil==2.5.3
pytz==2016.4
six==1.10.0
sqlparse==0.1.19
whichcraft==0.4.0
